* Implemented track draft local data source, * Settled on dark theme-only for HawkSpeed, * Implemented the track racing view model & all supporting functionality, * Wrote instrumented tests for track recording & track racing to prove the logic works, * Updated changelog to reflect latest details, * Added track types, * Added maps style, * Added various icons, * Progressed overall UI, * Updated README.

+++ data/src/main/java/com/vljx/hawkspeed/data/models/world/PlayerUpdateResultModel.kt
@@ -3,6 +3,6 @@ package com.vljx.hawkspeed.data.models.world
 data class PlayerUpdateResultModel(
     val latitude: Double,
     val longitude: Double,
-    val rotation: Double,
+    val rotation: Float,
     val worldObjectUpdateResult: WorldObjectUpdateResultModel?
 )
