+++ domain/src/main/java/com/vljx/hawkspeed/domain/models/world/PlayerUpdateResult.kt
@@ -3,6 +3,6 @@ package com.vljx.hawkspeed.domain.models.world
 data class PlayerUpdateResult(
     val latitude: Double,
     val longitude: Double,
-    val rotation: Double,
+    val rotation: Float,
     val worldObjectUpdateResult: WorldObjectUpdateResult?
 )
